ChessFlavour: Add code for chess like game
Added code for a chess like game with slightly different requirements
and a lot less features than an actual chess game.
Please don't use this code to learn programing a chess game. This is a
different flavour which can be used to learn about OOP and a few
concepts of chess programing.

Signed-off-by: Kamal Pandey <pandeykamal13526@gmail.com>